echonet

Module protocol

messaging packet encoder and decoder module.

Structs

Message

Message represents a messaging packet between ECHONET-Lite nodes as specified in the ECHONET-Lite specification.

Property

Each ECHONET-Lite message has properties. Property represents an ECHONET-Lite property of a message packet as specified in the ECHONET-Lite specification.

Enums

ESV

ESV represents an ECHONET-Lite service (ESV) code as specified in the ECHONET-Lite specification.

Constants

TID_MAX

TID_MIN

Type Aliases

PropertyCode

PropertyCode represents an ECHONET-Lite property code (EPC) in an ECHONET-Lite message.

Result

A Result type for protocol decoding and encoding operations.

TID

TID represents an ECHONET-Lite transaction identification number as specified in the ECHONET-Lite specification.