echonet

Struct Device

pub struct Device { /* private fields */ }

Device represents an ECHONET-Lite device node. The device is created with a standard device object code, and the object is automatically added the standard properties which are specified in APPENDIX Detailed Requirements for ECHONET Device Objects by the ECHONET Consortium.

Examples

use std::sync::{Arc, Mutex};
use echonet::protocol::{ESV, Property};
use echonet::util::Bytes;
use echonet::{Device, Object, ObjectCode, RequestHandler};

pub struct MyDevice {
    device: Device,
    on: bool,
}

impl MyDevice {
    pub fn new() -> Arc<Mutex<MyDevice>> {
        let m = Arc::new(Mutex::new(MyDevice {
            device: Device::new(0x029101),
            on: false,
        }));
        m.lock().unwrap().device.set_request_handler(m.clone());
        m
    }

    pub fn turn_on(&mut self) {
        self.on = true;
    }

    pub fn turn_off(&mut self) {
        self.on = false;
    }
}

impl RequestHandler for MyDevice {
    fn property_request_received(&mut self, deoj: &mut Object, esv: ESV, prop: &Property) -> bool {
        // Ignore all messages to other objects in the same node.
        if deoj.code() != self.device.code() {
            return false;
        }

        match esv {
            ESV::WriteRequest | ESV::WriteReadRequest => {
                let prop_code = prop.code();
                let prop_bytes = prop.data();
                match prop_code {
                    0x80 /* Operating status */ => {
                        let prop_u32 = Bytes::to_u32(prop_bytes);
                        match prop_u32 {
                            0x30 /* On */=> {
                                self.turn_on();
                                return true;
                            }
                            0x31 /* Off */=> {
                                self.turn_off();
                                return true;
                            }
                            _ => {}
                        }
                    }
                    _ => {}
                }
            }
            _ => {}
        }
        false
    }
}

Implementations

impl Device

pub fn new(code: ObjectCode) -> Device

Creates a new device with a specified object code. The object is automatically added the standard properties which are specified in APPENDIX Detailed Requirements for ECHONET Device Objects by the ECHONET Consortium.

pub fn new_with_node(code: ObjectCode, node: Arc<Mutex<Node>>) -> Device

Creates a new device with a specified object code and node to which it belongs. The functions is used when a node has multiple devices.

pub fn code(&self) -> ObjectCode

Returns the object code.

pub fn node(&self) -> Arc<Mutex<Node>>

Returns the parent local node to which the device belongs.

pub fn set_property(&mut self, code: u8, data: &[u8]) -> bool

Sets the data into the specified property if the device node has the property, otherwise return false.

pub fn property(&self, code: u8) -> Option<Vec<u8>>

Gets the specified property data if the device node has the property, otherwise return none.

pub fn set_request_handler(&mut self, handler: RequestHandlerObject)

pub fn start(&mut self) -> bool

Starts the device to communicate with other ECHONET-Lite nodes on the local network.

pub fn stop(&mut self) -> bool

Stops the device.

impl Device

pub fn set_manufacturer_code(&mut self, code: u32) -> bool

pub fn set_operating_status(&mut self, status: bool) -> bool

pub fn set_installation_location(&mut self, loc: u8) -> bool

pub fn set_standard_version(&mut self, ver: u8) -> bool

pub fn standard_version(&mut self) -> Option<Vec<u8>>

pub fn set_fault_status(&mut self, status: bool) -> bool