Crate echonet

echonet (uecho-rs)

The uecho-rs is a portable and cross-platform development framework for creating controller applications and devices of ECHONET-Lite for Rust developers. ECHONET-Lite is an open standard specification for IoT devices in Japan, it specifies more than 100 IoT devices such as crime prevention sensors, air conditioners and refrigerators.

What is uEcho ?

To implement IoT controllers or devices of ECHONET-Lite, the developer had to understand and implement the communication middleware specification such as the message format and base sequences.

The uecho-rs supports to control devices of ECHONET-Lite or create the standard devices of the specification easily. Using the uecho-rs, the developer has only to set basic listeners to implement the devices and controllers because the uecho-rs handles other requests such as read and notification requests automatically.

Getting Started

The examples are available for ECHONET-Lite controller and device implementations using the uecho-rs, check out the examples folder in GitHub.

Getting Help

Generated Docs (latest version)

Usage examples

License

This project is licensed under the Apache-2.0 License.

Modules

log

Logger function module.

protocol

Messaging protocol encoder and decoder module.

transport

Messaging transport manager module (Internal).

util

Utility function module.

Structs

Controller

Controller represents an ECHONET-Lite controller node to communicate other ECHONET-Lite nodes.

Device

Device represents an ECHONET-Lite device node.

LocalNode

LocalNode represents an ECHONET-Lite node in the controller and device nodes.

Manufacture

Manufacture represents a manufacturer name and code registerd by the ECHONET CONSORTIUM.

Object

Each ECHONET-Lite node has objects. Object represents an ECHONET-Lite Object in an ECHONET-Lite node.

Property

Each ECHONET-Lite object has properties. Property represents an ECHONET-Lite property in an ECHONET-Lite object.

PropertyEnum

PropertyEnum represents an ECHONET-Lite property enumerated data for an ECHONET-Lite property.

RemoteNode

RemoteNode represents an ECHONET-Lite node discovered by Controller. The remote node has the standard objects and properties defined by the ECHONET CONSORTIUM.

StandardDatabase

StandardDatabase represents a standard database for official device objects and manufactures defined by the ECHONET CONSORTIUM.

Enums

PropertyRule

PropertyRule represents an ECHONET-Lite property access rule for an ECHONET-Lite property.

Type Definitions

ManufactureCode

ManufactureCode represents a manufacturer code registerd by the ECHONET CONSORTIUM.